Cloud Function_ Automatización del ETL/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==1.28.0
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0